cleanup from review comments
