+++ browser/tor/tor_profile_service.cc
@@ -7,11 +7,17 @@
 
 #include <string>
 
+#include "brave/browser/brave_browser_process_impl.h"
+// TODO(bridiver) - move this out of extensions
+#include "brave/browser/extensions/brave_tor_client_updater.h"
 #include "brave/browser/tor/tor_launcher_service_observer.h"
 #include "brave/common/tor/pref_names.h"
-#include "brave/common/tor/tor_proxy_uri_helper.h"
+#include "chrome/common/channel_info.h"
+#include "components/version_info/channel.h"
 #include "components/pref_registry/pref_registry_syncable.h"
 #include "components/prefs/pref_registry_simple.h"
+#include "components/prefs/pref_service.h"
+
 
 namespace tor {
 
@@ -26,9 +32,41 @@ void TorProfileService::RegisterProfilePrefs(
     user_prefs::PrefRegistrySyncable* registry) {
   registry->RegisterBooleanPref(tor::prefs::kProfileUsingTor, false);
 }
+
 // static
 void TorProfileService::RegisterLocalStatePrefs(PrefRegistrySimple* registry) {
-  registry->RegisterStringPref(tor::prefs::kTorProxyString, GetTorProxyURI());
+  const std::string TorProxyScheme("socks5://");
+  const std::string TorProxyAddress("127.0.0.1");
+  std::string port;
+  switch (chrome::GetChannel()) {
+    case version_info::Channel::STABLE:
+      port = std::string("9350");
+      break;
+    case version_info::Channel::BETA:
+      port = std::string("9360");
+      break;
+    case version_info::Channel::DEV:
+      port = std::string("9370");
+      break;
+    case version_info::Channel::CANARY:
+      port = std::string("9380");
+      break;
+    case version_info::Channel::UNKNOWN:
+    default:
+      port = std::string("9390");
+  }
+  const std::string tor_proxy_uri =
+      std::string(TorProxyScheme + TorProxyAddress + ":" + port);
+  registry->RegisterStringPref(tor::prefs::kTorProxyString, tor_proxy_uri);
+}
+
+std::string TorProfileService::GetTorProxyURI() {
+  return g_browser_process->local_state()->GetString(
+      tor::prefs::kTorProxyString);
+}
+
+base::FilePath TorProfileService::GetTorExecutablePath() {
+  return g_brave_browser_process->tor_client_updater()->GetExecutablePath();
 }
 
 void TorProfileService::AddObserver(TorLauncherServiceObserver* observer) {
