+++ browser/tor/tor_profile_service.h
@@ -51,6 +51,8 @@ class TorProfileService : public KeyedService {
   void RemoveObserver(TorLauncherServiceObserver* observer);
 
  protected:
+  std::string GetTorProxyURI();
+  base::FilePath GetTorExecutablePath();
   base::ObserverList<TorLauncherServiceObserver> observers_;
 
  private:
