--- common/tor/tor_proxy_uri_helper.h
@@ -1,16 +1,0 @@
-/* Copyright (c) 2019 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#ifndef BRAVE_COMMON_TOR_TOR_PROXY_URI_HELPER_H_
-#define BRAVE_COMMON_TOR_TOR_PROXY_URI_HELPER_H_
-
-#include <string>
-
-namespace tor {
-
-std::string GetTorProxyURI();
-
-}  // namespace tor
-#endif  // BRAVE_COMMON_TOR_TOR_PROXY_URI_HELPER_H_
